// frontend/src/api/generated/models/materialResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface MaterialResponse {
  id: string;
  code: string;
  name: string;
  purchase_uom_id: string;
  purchase_uom_code: string;
  usage_uom_id: string;
  usage_uom_code: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  conversion_ratio: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  moq: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  purchase_multiple: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  package_quantity: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  shrinkage_percent: string;
  default_supplier_id?: string | null;
  last_purchase_price?: string | null;
  is_active: boolean;
}